requirements.txt
django-allauth
psycopg2
django
